update: move FileStat to src/lib
