+++ src/lib/file-stat.ts
